Move PWM into the HAL.
